Write and modify sorting algo tests and create package for classes
