+++ src/calebowusuyianoma/sortalgovisualiser/ArrayGenerator.java
@@ -1,3 +1,5 @@
+package calebowusuyianoma.sortalgovisualiser;
+
 import java.util.ArrayList;
 
 public class ArrayGenerator {
