+++ src/calebowusuyianoma/sortalgovisualiser/Main.java
@@ -1,3 +1,5 @@
+package calebowusuyianoma.sortalgovisualiser;
+
 public class Main {
 
     public static void main(String[] args) {
