+++ src/calebowusuyianoma/sortalgovisualiser/MainPanel.java
@@ -1,3 +1,5 @@
+package calebowusuyianoma.sortalgovisualiser;
+
 import javax.swing.JButton;
 import javax.swing.JPanel;
 import javax.swing.Timer;
