+++ src/calebowusuyianoma/sortalgovisualiser/MergeSort.java
@@ -1,0 +1,37 @@
+package calebowusuyianoma.sortalgovisualiser;
+
+import java.util.ArrayList;
+
+public class MergeSort {
+    public void sort(ArrayList<Integer> data) {
+        sort(data, 0, data.size() - 1);
+    }
+
+    private void sort(ArrayList<Integer> data, int lowIndex, int highIndex) {
+        if(lowIndex < highIndex) {
+            int middleIndex = (int)Math.floor((double)(lowIndex + highIndex) / 2);
+            sort(data, lowIndex, middleIndex);
+            sort(data, middleIndex + 1, highIndex);
+            merge(data, lowIndex, middleIndex, highIndex);
+        }
+    }
+
+    private void merge(ArrayList<Integer> data, int lowIndex, int middleIndex, int highIndex) {
+        ArrayList<Integer> leftSubArray = new ArrayList<>(data.subList(lowIndex, middleIndex + 1));
+        ArrayList<Integer> rightSubArray = new ArrayList<>(data.subList(middleIndex + 1, highIndex + 1));
+
+        leftSubArray.add(Integer.MAX_VALUE);
+        rightSubArray.add(Integer.MAX_VALUE);
+
+        int i = 0, j = 0;
+        for(int k = lowIndex; k < highIndex + 1; k++) {
+            if(leftSubArray.get(i) < rightSubArray.get(j)) {
+                data.set(k, leftSubArray.get(i));
+                i++;
+            } else {
+                data.set(k, rightSubArray.get(j));
+                j++;
+            }
+        }
+    }
+}
